+++ src/Twig/AppExtension.php
@@ -7,7 +7,7 @@
 namespace WizaplaceFrontBundle\Twig;
 
 use Symfony\Component\Asset\Packages;
-use Wizaplace\SDK\Catalog\CatalogService;
+use Wizaplace\SDK\Catalog\CatalogServiceInterface;
 use Wizaplace\SDK\Cms\CmsService;
 use Wizaplace\SDK\Image\Image;
 use Wizaplace\SDK\Image\ImageService;
@@ -18,7 +18,7 @@ use WizaplaceFrontBundle\Service\ProductUrlGenerator;
 
 class AppExtension extends \Twig_Extension
 {
-    /** @var CatalogService */
+    /** @var CatalogServiceInterface */
     private $catalogService;
     /** @var BasketService */
     private $basketService;
@@ -38,7 +38,7 @@ class AppExtension extends \Twig_Extension
     private $favoriteService;
 
     public function __construct(
-        CatalogService $catalogService,
+        CatalogServiceInterface $catalogService,
         BasketService $basketService,
         ImageService $imageService,
         CmsService $cmsService,
@@ -100,9 +100,6 @@ class AppExtension extends \Twig_Extension
         return (string) $this->imageService->getImageLink($imageId, $width, $height);
     }
 
-    /**
-     * @throws \Psr\Cache\InvalidArgumentException
-     */
     public function getCategoryTree():array
     {
         return $this->catalogService->getCategoryTree();